Fix install + uninstall (#84)

+++ uninstall.sh
@@ -1,13 +1,25 @@
 #!/usr/bin/env bash
 
 # Nodeyez uninstall script
+DELETED_SERVICES=0
+DELETED_ENV=0
 DELETED_WEBSITE=0
+DELETED_UFWRULES=0
+DELETED_SELFCERT=0
 DELETED_USER=0
 
 # Remove Nodeyez services
 for f in $(systemctl list-unit-files --type=service | grep nodeyez | awk '{print $1}'); do sudo systemctl stop $f; sudo systemctl disable $f; done
-sudo rm /etc/systemd/system/nodeyez-*.service
-sudo rm /etc/nodeyez.conf
+# - service files
+if [ $(sudo ls -la /etc/systemd/system | grep nodeyez | wc -l) -gt 0 ]; then
+  sudo rm /etc/systemd/system/nodeyez-*.service
+  DELETED_SERVICES=1
+fi
+# - environment file
+if [ -f "/etc/nodeyez.conf" ]; then
+  sudo rm /etc/nodeyez.conf
+  DELETED_ENV=1
+fi
 
 # Remove Nodeyez dashboard
 if [ $(which nginx | wc -l) -gt 0 ]; then
@@ -42,14 +54,19 @@ if [ $(which nginx | wc -l) -gt 0 ]; then
 fi
 
 # Remove uncomplicated firewall rules with Nodeyez comments
-for n in $(sudo ufw status numbered | grep Nodeyez | awk '{print $2}' | sed 's/]//' | sort -r); do sudo ufw --force delete $n; done
+if [ $(sudo ufw status | grep Nodeyez | wc -l) -gt 0 ]; then
+  for n in $(sudo ufw status numbered | grep Nodeyez | awk '{print $2}' | sed 's/]//' | sort -r); do sudo ufw --force delete $n; done
+  DELETED_UFWRULES=1
+fi
 
 # Remove Nodeyez self-signed cert
 if [ -f "/etc/ssl/certs/nodeyez-nginx-selfsigned.crt" ]; then
   sudo rm /etc/ssl/certs/nodeyez-nginx-selfsigned.crt
+  DELETED_SELFCERT=1
 fi
 if [ -f "/etc/ssl/private/nodeyez-nginx-selfsigned.key" ]; then
   sudo rm /etc/ssl/private/nodeyez-nginx-selfsigned.key
+  DELETED_SELFCERT=1
 fi
 
 
@@ -73,12 +90,27 @@ fi
 # show summary
 echo "================================================"
 echo "SUMMARY:"
-echo "Remaining services"
-sudo systemctl list-units --type=service --state=active | grep nodeyez | awk '{print "(active) " $1}'
-sudo systemctl list-units --type=service --state=failed | grep nodeyez | awk '{print "(failed) " $2}'
+if [ $DELETED_SERVICES -ge 1 ]; then
+  echo "Nodeyez services removed"
+fi
+if [ $(sudo systemctl list-units --type=service | grep nodeyez | wc -l) -ge 1 ]; then
+  echo "Remaining services"
+  sudo systemctl list-units --type=service --state=active | grep nodeyez | awk '{print "(active) " $1}'
+  sudo systemctl list-units --type=service --state=failed | grep nodeyez | awk '{print "(failed) " $2}'
+fi
+if [ $DELETED_ENV -ge 1 ]; then
+  echo "Nodeyez environment variables for services removed"
+fi
 if [ $DELETED_WEBSITE -ge 1 ]; then
   echo "Website deleted"
+fi
+if [ $DELETED_UFWRULES -ge 1 ]; then
+  echo "Uncomplicated Firewall Rules for Nodeyez removed"
+fi
+if [ $DELETED_SELFCERT -ge 1 ]; then
+  echo "Nodeyez self-signed certificates removed"
 fi
 if [ $DELETED_USER -ge 1 ]; then
   echo "Nodeyez user deleted"
 fi
+echo "Uninstallation complete"
